updating some metadata handling
